Create LICENSE.TXT
Added License file, hopefully it's the right one. I am not a lawyer. 
But this project, due to it's use of CC-CEDIT, is under the "Creative Commons Attribution-ShareAlike 4.0 International License". To quote the CC-CEDICT website: "It more or less means that you are allowed to use this data for both non-commercial and commercial purposes provided that you: mention where you got the data from (attribution) and that in case you improve / add to the data you will share these changes under the same license (share alike)."