give up
This is as far as I got.
Still has issues with things like finding multiple entries.
For example, we are just guessing which dictionary entry is good enough in one instance (like 96), and there was still an issue in the end with an alternate form of 比 showing up like 3 different pinyin versions