Create mnemonics.u8
New sideproject: Mnemonic help.
Hanzi can be hard, so a seperate file is started for mnemonics only.

Format still is vaguely u8-like, but not really following the same style. Because explainations need proper grammar, styling etc, everything will be 'one definition', will be capitalized properly etc.
If for some reason a linebreak is necessary (for example, an alternative way to remember a word), then a forward slash may be used